requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.1
blis==0.7.8
boto3==1.24.39
botocore==1.27.39
bs4==0.0.1
cachelib==0.9.0
cachetools==5.2.0
catalogue==1.0.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.32
cytoolz==0.12.0
emoji==1.7.0
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.7.1
Flask==2.1.3
Flask-Session==0.4.0
fonttools==4.34.4
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
murmurhash==1.0.7
networkx==2.6.3
nltk==3.7
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
plac==1.1.3
pluggy==1.0.0
preshed==3.0.6
py==1.11.0
pyemd==0.5.1
pyparsing==3.0.9
pyphen==0.12.0
pyScss==1.4.0
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.25
requests==2.28.0
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.3.2.post1
spacy==2.3.7
# neuralcoref==4.0 --no-binary==neuralcoref
srsly==1.0.5
textacy==0.10.0
thinc==7.4.5
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
toolz==0.12.0
torch==1.12.0
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.3.0
urllib3==1.26.9
wasabi==0.10.1
Werkzeug==2.2.1
zipp==3.8.1
gensim==4.2.0
flask_socketio==5.2.0